refactor: move helpers to SetupStorage.kt
Prep for moving CollectionHelper to :anki-common.

Collection creation will become a startup concern, but collection
 access will move to :anki-common.

* Part of 19552
* Part of 13574
* Part of 20737
